+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Upcoming
 * Support regular tag parameters (`[tag param=value][/tag]`) instead of 'quick parameters' (`[tag=value][/tag]`)
 * Changed tag description symbols (to become descriptive), **breaks existing custom tag additiona!**
 * Add support to show the BBCode annotated with errors (when there are any)
+* Add support to escape token value using :uri_escape.
 
 Version 1.0.1 - 04-Jan-2015